Rename sample to simplify name
