+++ Photino.Blazor.sln
@@ -17,7 +17,7 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution 
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Samples", "Samples", "{1F04F43F-CA2D-44E4-8984-4B8E4B875BD3}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Photino.Blazor.Samples.HelloWorld", "Samples\1. Hello World\Photino.Blazor.Samples.HelloWorld.csproj", "{B5A5E655-DCAE-4553-AC8B-6E26287D9C3B}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "_1_Hello_World", "Samples\_1_Hello_World\_1_Hello_World.csproj", "{B5A5E655-DCAE-4553-AC8B-6E26287D9C3B}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
